fix(development): fix bug in transfer server list

+++ src/db/include/awsmock/repository/TransferDatabase.h
@@ -166,7 +166,7 @@ namespace AwsMock::Database {
          * @param sortColumns sorting column names
          * @return list of transfer users
          */
-        std::vector<Entity::Transfer::User> ListUsers(const std::string &region, const std::string &serverId, const std::string &prefix = {}, int pageSize = 0, int pageIndex = 0, const std::vector<SortColumn> &sortColumns = {}) const;
+        std::vector<Entity::Transfer::User> ListUsers(const std::string &region, const std::string &serverId, const std::string &prefix = {}, long pageSize = 0, long pageIndex = 0, const std::vector<SortColumn> &sortColumns = {}) const;
 
         /**
          * @brief Returns a list of transfer users.
